+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/IrOnlyClassesBoxTestGenerated.java
@@ -3633,6 +3633,12 @@ public class IrOnlyClassesBoxTestGenerated extends AbstractIrOnlyClassesBoxTest 
     }
 
     @Test
+    @TestMetadata("NestedGraphExtensionMiddleScopedBinding.kt")
+    public void testNestedGraphExtensionMiddleScopedBinding() {
+      run("NestedGraphExtensionMiddleScopedBinding.kt");
+    }
+
+    @Test
     @TestMetadata("NestedSuspendFunctionWrappersWithoutRuntime.kt")
     public void testNestedSuspendFunctionWrappersWithoutRuntime() {
       run("NestedSuspendFunctionWrappersWithoutRuntime.kt");