+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/JsFastInitBoxTestGenerated.java
@@ -2805,6 +2805,12 @@ public class JsFastInitBoxTestGenerated extends AbstractJsFastInitBoxTest {
     }
 
     @Test
+    @TestMetadata("NestedGraphExtensionMiddleScopedBinding.kt")
+    public void testNestedGraphExtensionMiddleScopedBinding() {
+      run("NestedGraphExtensionMiddleScopedBinding.kt");
+    }
+
+    @Test
     @TestMetadata("NestedSuspendFunctionWrappersWithoutRuntime.kt")
     public void testNestedSuspendFunctionWrappersWithoutRuntime() {
       run("NestedSuspendFunctionWrappersWithoutRuntime.kt");